Attempt 3 - The new store json from webapi is now where it belongs
